Add `pt-pt` alias for `pt` language flag
Since Brazilian Portuguese is the globally dominant variety of Portuguese
(in terms of number of speakers and cultural influence)
similar to how American English is for English,
it is common for people to use a `pt-pt` language code
to specifically refer to European Portuguese.

This commit therefore adds an alias `pt-pt` -> `pt`,
similar to the existing `en-gb` -> `en` alias for English (United Kingdom).
Despite the apparent redundancy, the `pt-pt` code is actually used in practice,
and indeed has an IETF language tag (`pt-PT`) and POSIX locale (`pt_PT`).